requirements.txt
kafka-python
pandas
numpy
scikit-learn
fastapi
uvicorn
streamlit

